1.3.4

+++ command/bigfun.js
@@ -2,8 +2,12 @@ const Discord = require("discord.js");
 
 module.exports.run = async (client, message, args, tools) => {
    
+    let bigfun = message.guild.roles.find("name", "• bigfun •");
+    if(!message.member.roles.has(bigfunRole.id)) return message.member.send(`NO PORN FOR YOU!!!`);
+    if(message.member.roles.has(bigfunRole.id)){
+
     message.member.send("https://rule34.xxx/index.php?page=post&s=view&id=3312363");
 
     message.delete().catch(O_o=>{});
-  
+    }
 }
